% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chron.R
\name{chron_context}
\alias{chron_context}
\title{Print Chronicling America context for LLM integration}
\usage{
chron_context()
}
\value{
Invisibly returns the context string
}
\description{
Print Chronicling America context for LLM integration
}
